Splitting the expansion into files, started using compile_flags.txt, starting to implement javadoc style commenting.
