+++ compile_flags.txt
@@ -1,0 +1,4 @@
+-I
+include/ 
+-I
+libft/
