+++ tests/jvarila/minishell_test_word_lexing_01.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/minishell.h"
+#include "minishell.h"
 
 void	add_token_to_end(t_token **list, t_token *token)
 {
@@ -61,9 +61,8 @@ void	insert_token_left(t_token *current, t_token *new)
 }
 
 /**
- * DESCRIPTION
- *     The initial token count treats "|<> \t\n" as separating characters,
- *     later words will be expanded and split which will impact the count.
+ * The initial token count treats "|<> \t\n" as separating characters,
+ * later words will be expanded and split which will impact the count.
  */
 size_t	count_tokens(const char *str)
 {
@@ -197,53 +196,6 @@ void	print_tokens(t_minishell *data)
 	}
 }
 
-t_var	*find_var(t_minishell *data, const char *str)
-{
-	t_var	*var;
-	size_t	len;
-
-	len = 0;
-	while (str[len] && !ft_isspace(str[len]) \
-		&& !ft_strchr(EXPANSION_DELIMITER, str[len]))
-		len++;
-	var = data->custom_env;
-	while (var)
-	{
-		if (ft_strncmp(var->key, str, len) == 0)
-			return (var);
-		var = var->next;
-	}
-	return (NULL);
-}
-
-void	expand_variable(t_minishell *data, t_token *token, \
-					  t_var *var, const char *var_pos)
-{
-	char	*new;
-	size_t	old_len;
-	size_t	new_len;
-	size_t	var_key_len;
-	size_t	var_value_len;
-
-	if (!var)
-	{
-		var_key_len = ft_strlen(var_pos);
-		var_value_len = 0;
-	}
-	else
-	{
-		var_key_len = ft_strlen(var->key);
-		var_value_len = ft_strlen(var->value);
-	}
-	old_len = ft_strlen(token->value);
-	new_len = old_len - 1 - var_key_len + var_value_len;
-	new = memarena_calloc(data->arena, new_len + 1, sizeof(char));
-	ft_strlcat(new, token->value, var_pos - token->value);
-	if (var)
-		ft_strlcat(new, var->value, new_len + 1);
-	ft_strlcat(new, var_pos + var_key_len, new_len + 1);
-	token->value = new;
-}
 
 void	toggle_bool(bool *value)
 {
@@ -253,68 +205,6 @@ void	toggle_bool(bool *value)
 		*value = true;
 }
 
-bool	contains_unexpanded_variable(t_token *token)
-{
-	const char	*str;
-	bool		within_single_quotes;
-
-	str = token->value;
-	within_single_quotes = false;
-	while (str)
-	{
-		str = skip_whitespace(str);
-		if (!*str)
-			break ;
-		if (*str++ == '\'')
-		{
-			toggle_bool(&within_single_quotes);
-			continue ;
-		}
-		if (*str == '$' && *(str + 1) && !ft_isspace(*(str + 1)) && !within_single_quotes)
-			return (true);
-		str++;
-	}
-	return (false);
-}
-
-void	expand_variables(t_minishell *data, t_token *token)
-{
-	char	*str;
-	t_var	*var;
-
-	str = ft_strchr(token->value, '$');
-	while (contains_unexpanded_variable(token))
-	{
-		str++;
-		if (ft_isspace(*str))
-		{
-			str = ft_strchr(str, '$');
-			continue ;
-		}
-		var = find_var(data, str);
-		expand_variable(data, token, var, str);
-		str = ft_strchr(token->value, '$');
-	}
-}
-
-void	variable_expansion(t_minishell *data)
-{
-	char	*str;
-	t_token	*token;
-
-	token = data->token_list;
-	while (token)
-	{
-		str = ft_strchr(token->value, '$');
-		if (!str)
-		{
-			token = token->next;
-			continue ;
-		}
-		expand_variables(data, token);
-		token = token->next;
-	}
-}
 
 int	main(int argc, char *argv[])
 {
